add seo and router

+++ protected/controllers/admin/ArticleController.php
@@ -27,27 +27,21 @@ class ArticleController extends AdminController
     }    
     
     protected function updateRelData($model) {
-        $tagStr = '';
-        if (isset($_POST['tags'])) {
-            $tagStr = $_POST['tags'];
-        }
-        
-        $tags = explode(';', $tagStr);
-        
-        Tag::deleteAllRelations($model->id);
-        if ($tags) foreach ($tags as $tagName){
-            $tag = Tag::get($tagName);
-            $tag->addRelation($model->id);
+        if (isset($_POST['tags_dirty']) && ($_POST['tags_dirty'] == 1) ){
+            $tagStr = '';
+            if (isset($_POST['tags'])) {
+                $tagStr = $_POST['tags'];
+            }
+
+            $tags = explode(';', $tagStr);
+
+            Tag::deleteAllRelations($model->id);
+            if ($tags) foreach ($tags as $tagName){
+                $tag = Tag::get($tagName);
+                $tag->addRelation($model->id);
+            }
         }
         return true;
     }
     
-    protected function processData($data) {
-        if (isset($data['published'])){
-            //19 07 2014 - 10:30 => 2014-07-19 10:30:00
-            $publ = explode(' ', $data['published']);
-            $data['published'] = $publ[2].'-'.$publ[1].'-'.$publ[0].' '.$publ[4].':00';
-        }
-        return $data;
-    }
 }
